Extract PasswordHasher.ComparisonResult to its own class
