+++ app/src/main/java/org/simple/clinic/login/applock/BCryptPasswordHasher.kt
@@ -2,13 +2,13 @@ package org.simple.clinic.login.applock
 
 import io.reactivex.Single
 import org.mindrot.jbcrypt.BCrypt
-import org.simple.clinic.login.applock.PasswordHasher.ComparisonResult.DIFFERENT
-import org.simple.clinic.login.applock.PasswordHasher.ComparisonResult.SAME
+import org.simple.clinic.login.applock.ComparisonResult.DIFFERENT
+import org.simple.clinic.login.applock.ComparisonResult.SAME
 import javax.inject.Inject
 
 class BCryptPasswordHasher @Inject constructor() : PasswordHasher {
 
-  override fun compare(hashed: String, password: String): Single<PasswordHasher.ComparisonResult> {
+  override fun compare(hashed: String, password: String): Single<ComparisonResult> {
     return Single.fromCallable {
       val checkpw = BCrypt.checkpw(password, hashed)
       when (checkpw) {
