+++ app/src/main/java/org/simple/clinic/login/applock/ComparisonResult.kt
@@ -1,0 +1,6 @@
+package org.simple.clinic.login.applock
+
+enum class ComparisonResult {
+  SAME,
+  DIFFERENT
+}
