+++ app/src/main/java/org/simple/clinic/login/applock/PasswordHasher.kt
@@ -4,11 +4,6 @@ import io.reactivex.Single
 
 interface PasswordHasher {
 
-  enum class ComparisonResult {
-    SAME,
-    DIFFERENT
-  }
-
   fun hash(password: String): Single<String>
 
   fun compare(hashed: String, password: String): Single<ComparisonResult>
